requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
black==24.4.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.6.0
grandalf==0.8
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
isort==5.13.2
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.17
langchain-community==0.0.38
langchain-core==0.1.53
langchain-openai==0.1.6
langchain-text-splitters==0.0.2
langchainhub==0.1.15
langgraph==0.0.44
langsmith==0.1.147
marshmallow==3.26.1
multidict==6.4.3
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.75.0
orjson==3.10.16
packaging==23.2
pathspec==0.12.1
platformdirs==4.3.7
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tenacity==8.5.0
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
